+++ docs/search.json
@@ -35,11 +35,11 @@
     "text": "Visualization 2: [Title Here]\nDescribe what this plot shows, and why it’s important.\n\n\nCode\n# your code here"
   },
   {
-    "objectID": "index.html#visualization-3-title-here",
-    "href": "index.html#visualization-3-title-here",
+    "objectID": "index.html#visualization-3-does-sex-education-type-affect-the-type-of-contraception-used",
+    "href": "index.html#visualization-3-does-sex-education-type-affect-the-type-of-contraception-used",
     "title": "Final Project: [YOUR TITLE HERE]",
-    "section": "Visualization 3: [Title Here]",
-    "text": "Visualization 3: [Title Here]\nDescribe what this plot shows, and why it’s important.\n\n\nCode\n# your code here"
+    "section": "Visualization 3: [Does sex education type affect the type of contraception used?]",
+    "text": "Visualization 3: [Does sex education type affect the type of contraception used?]\nThis question examines whether the type of sex education individuals received influences their current contraceptive behavior. Using the variable abstonly, we categorize respondents by whether they received abstinence-focused education, comprehensive sex education, or no sex education. We then analyze their reported contraceptive method. The goal is to identify whether sex education is associated with the use of more effective or consistent contraceptive methods, offering insight into how educational content may shape real-life health decisions.\n\n\nCode\nlibrary(tidyverse)\nlibrary(ggalluvial)\n\nd &lt;- read_csv(\"clean_fogzone.csv\")\n\nsankey_data &lt;- d %&gt;%\n  select(abstonly, Current_most_effective_method) %&gt;%\n  filter(\n    !is.na(abstonly),\n    !is.na(Current_most_effective_method),\n    !str_detect(str_to_lower(Current_most_effective_method), \"no sex|refused|don't know\")\n  ) %&gt;%\n  count(abstonly, Current_most_effective_method) %&gt;%\n  rename(sex_ed = abstonly, method = Current_most_effective_method) %&gt;%\n  mutate(\n    sex_ed = case_when(\n      sex_ed == \"Abstinence Focused\" ~ \"Abstinence-Focused\",\n      sex_ed == \"Abstinence Only\" ~ \"Abstinence Only\",\n      sex_ed == \"Contraceptive Focused\" ~ \"Comprehensive\",\n      sex_ed == \"No sex ed/ don't know\" ~ \"No/Don't Know\",\n      TRUE ~ sex_ed\n    ),\n    method = case_when(\n      method == \"hormonal + condoms\" ~ \"Hormonal + Condoms\",\n      method == \"hormonal alone\" ~ \"Hormonal Alone\",\n      method == \"long-acting alone\" ~ \"LARC Alone\",\n      method == \"long-acting + condom\" ~ \"LARC + Condom\",\n      method == \"no method - likely at risk\" ~ \"No Method (Likely Risk)\",\n      method == \"no method - risk unknown\" ~ \"No Method (Unknown Risk)\",\n      method == \"other/unknown\" ~ \"Other/Unknown\",\n      method == \"withdrawal/NFP\" ~ \"Withdrawal / NFP\",\n      TRUE ~ method\n    )\n  )\n\np &lt;- ggplot(sankey_data,\n            aes(axis1 = sex_ed, axis2 = method, y = n)) +\n  geom_alluvium(aes(fill = sex_ed), width = 1/15) +\n  geom_stratum(width = 1/15, fill = \"orange\", color = \"black\") +\n  scale_x_discrete(\n    limits = c(\"Sex Education Type\", \"Contraceptive Method\"),\n    expand = c(.05, .05)\n  ) +\n  labs(\n    title = \"Sex Ed Type → Current Contraceptive Method\",\n    y = \"Number of People\",\n    fill = \"Sex Education\"\n  ) +\n  theme_classic()\n\n\nstrata_data &lt;- ggplot_build(p)$data[[2]]\n\np +\n  geom_text(data = strata_data %&gt;% filter(x == 1),\n            aes(x = x - 0.05, y = y, label = stratum),\n            hjust = 1, size = 3, inherit.aes = FALSE) +\n  geom_text(data = strata_data %&gt;% filter(x == 2),\n            aes(x = x + 0.05, y = y, label = stratum),\n            hjust = 0, size = 3, inherit.aes = FALSE)"
   },
   {
     "objectID": "index.html#or-shiny-app-title-here",
